requirements.txt
streamlit==1.28.1
openai==1.3.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
pandas==2.1.3
plotly==5.17.0